move faq-config to general-config
